backend/requirements.txt
google-genai>=1.0.0
google-cloud-firestore>=2.0.0
google-cloud-secret-manager>=2.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
aiohttp>=3.9.0
numpy>=1.26.0
sentence-transformers>=3.0.0
scipy>=1.12.0
